F_T30603 move DocumentBundle Transformer to core
- adjut use statement
