+++ demosplan/DemosPlanCoreBundle/Transformers/Document/DocumentDashboardTransformer.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanDocumentBundle\Transformers;
+namespace demosplan\DemosPlanCoreBundle\Transformers\Document;
 
 use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\Transformer\BaseTransformer;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\PlanningDocumentCategoryResourceType;
